@@ -1,7 +1,13 @@
-import React from "react";
-import AppBar from "material-ui/AppBar";
-import Drawer from "material-ui/Drawer";
-import MenuItem from "material-ui/MenuItem";
+import * as React from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import Typography from "@mui/material/Typography";
+import Button from "@mui/material/Button";
+import IconButton from "@mui/material/IconButton";
+import MenuIcon from "@mui/icons-material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import Drawer from "@mui/material/Drawer";
 import StatCalculator from "./StatCalculator.jsx";
 import InfoPage from "./InfoPage.jsx";
 
@@ -10,119 +16,111 @@ import g25 from "../res/json/gen2-5.mini.json";
 import g6 from "../res/json/gen6.mini.json";
 import g7 from "../res/json/gen7.mini.json";
 
-export default class PokresApp extends React.Component {
-  state = { data: g1, open: false, title: "Generation 1", generation: "1" };
+function getData(gen) {
+  if (gen === "1") {
+    return g1;
+  }
+  if (gen === "6") {
+    return g6;
+  }
+  if (gen === "7") {
+    return g7;
+  }
+  return g25;
+}
 
-  handleToggle = () => this.setState({ open: !this.state.open });
-  handleClose = () => this.setState({ open: false });
+export default function PokresApp() {
+  const [view, setView] = React.useState("default");
+  const [generation, setGeneration] = React.useState("1");
 
-  gen1 = () => {
-    this.setState({
-      open: false,
-      generation: "1",
-      title: "Generation 1",
-      data: g1,
-    });
+  const gen1 = () => {
+    setView("default");
+    setGeneration("1");
   };
 
-  gen2 = () => {
-    this.setState({
-      open: false,
-      generation: "2",
-      title: "Generation 2",
-      data: g25,
-    });
+  const gen2 = () => {
+    setView("default");
+    setGeneration("2");
   };
 
-  gen3 = () => {
-    this.setState({
-      open: false,
-      generation: "3",
-      title: "Generation 3",
-      data: g25,
-    });
+  const gen3 = () => {
+    setView("default");
+    setGeneration("3");
   };
 
-  gen4 = () => {
-    this.setState({
-      open: false,
-      generation: "4",
-      title: "Generation 4",
-      data: g25,
-    });
+  const gen4 = () => {
+    setView("default");
+    setGeneration("4");
   };
 
-  gen5 = () => {
-    this.setState({
-      open: false,
-      generation: "5",
-      title: "Generation 5",
-      data: g25,
-    });
+  const gen5 = () => {
+    setView("default");
+    setGeneration("5");
   };
 
-  gen6 = () => {
-    this.setState({
-      open: false,
-      generation: "6",
-      title: "Generation 6",
-      data: g6,
-    });
-  };
-  gen7 = () => {
-    this.setState({
-      open: false,
-      generation: "7",
-      title: "Generation 7",
-      data: g7,
-    });
-  };
-  info = () => {
-    this.setState({
-      open: false,
-      generation: "",
-      title: "Information",
-      data: null,
-    });
+  const gen6 = () => {
+    setView("default");
+    setGeneration("6");
   };
 
-  render() {
-    var mainContent;
-    if (this.state.data === null) {
-      mainContent = <InfoPage />;
-    } else {
-      mainContent = (
-        <StatCalculator
-          data={this.state.data}
-          generation={this.state.generation}
-        />
-      );
-    }
-    return (
-      <div>
-        <div className="nav_bar">
-          <AppBar
-            title={this.state.title}
-            onLeftIconButtonTouchTap={this.handleToggle}
-          />
-        </div>
-        <Drawer
-          docked={false}
-          width={200}
-          open={this.state.open}
-          onRequestChange={(open) => this.setState({ open })}
-        >
-          <MenuItem onClick={this.gen1}>Generation 1</MenuItem>
-          <MenuItem onClick={this.gen2}>Generation 2</MenuItem>
-          <MenuItem onClick={this.gen3}>Generation 3</MenuItem>
-          <MenuItem onClick={this.gen4}>Generation 4</MenuItem>
-          <MenuItem onClick={this.gen5}>Generation 5</MenuItem>
-          <MenuItem onClick={this.gen6}>Generation 6</MenuItem>
-          <MenuItem onClick={this.gen7}>Generation 7</MenuItem>
-          <MenuItem onClick={this.info}>Information</MenuItem>
-        </Drawer>
-        {mainContent}
-      </div>
+  const gen7 = () => {
+    setView("default");
+    setGeneration("7");
+  };
+
+  const info = () => {
+    setView("info");
+    setGeneration("7");
+  };
+
+  let mainContent;
+  if (view === "info") {
+    mainContent = <InfoPage />;
+  } else {
+    mainContent = (
+      <StatCalculator data={getData(generation)} generation={generation} />
     );
   }
+  return (
+    <div>
+      <div className="nav_bar">
+        <AppBar>
+          <Toolbar>
+            <IconButton
+              size="large"
+              edge="start"
+              color="inherit"
+              aria-label="menu"
+              sx={{ mr: 2 }}
+              onClick={() => {
+                if (view === "default") {
+                  setView("drawer");
+                } else {
+                  setView("default");
+                }
+              }}
+            >
+              <MenuIcon />
+            </IconButton>
+          </Toolbar>
+        </AppBar>
+      </div>
+      <Drawer
+        onClose={() => {
+          setView("default");
+        }}
+        open={view === "drawer"}
+      >
+        <MenuItem onClick={gen1}>Generation 1</MenuItem>
+        <MenuItem onClick={gen2}>Generation 2</MenuItem>
+        <MenuItem onClick={gen3}>Generation 3</MenuItem>
+        <MenuItem onClick={gen4}>Generation 4</MenuItem>
+        <MenuItem onClick={gen5}>Generation 5</MenuItem>
+        <MenuItem onClick={gen6}>Generation 6</MenuItem>
+        <MenuItem onClick={gen7}>Generation 7</MenuItem>
+        <MenuItem onClick={info}>Information</MenuItem>
+      </Drawer>
+      {mainContent}
+    </div>
+  );
 }
